+++ packages/components/organisms/f-checkout/src/services/tests/mapper.test.js
@@ -1,4 +1,4 @@
-import mapUpdateCheckoutRequest from '../mapper';
+import { mapUpdateCheckoutRequest, mapAnalyticsFieldNames } from '../mapper';
 
 const defaultParams = {
     address: {},
@@ -93,5 +93,74 @@ describe('checkout mapper', () => {
         // Assert
         expect(requestBody.notes.length).toBe(1);
         expect(requestBody.notes[0].note).toBe(userNote);
+    });
+});
+
+describe('mapAnalyticsFieldNames :: ', () => {
+    it.each([
+        ['address.line1', 'addressLine1'],
+        ['line1', 'addressLine1'],
+        ['address.line2', 'addressLine2'],
+        ['line2', 'addressLine2'],
+        ['address.city', 'addressCity'],
+        ['city', 'addressCity'],
+        ['address.postcode', 'addressPostcode'],
+        ['postcode', 'addressPostcode'],
+        ['customer.firstName', 'firstName'],
+        ['customer.lastName', 'lastName'],
+        ['customer.mobileNumber', 'phone'],
+        ['mobileNumber', 'phone'],
+        ['customer.email', 'email']
+    ])('should map the fieldname %s to the analytics value %s', (provided, expected) => {
+        // Act & Assert
+        expect(mapAnalyticsFieldNames(provided)).toEqual(expected);
+    });
+
+    it('should correctly map an array of field names and sort alphabetically', () => {
+        // Arrange
+        const provided = [
+            'city',
+            'address.line1',
+            'customer.email',
+            'customer.firstName',
+            'lastName',
+            'mobilePhone'
+        ];
+
+        const expected = [
+            'addressCity',
+            'addressLine1',
+            'email',
+            'firstName',
+            'lastName',
+            'mobilePhone'
+        ];
+
+        // Act & Assert
+        expect(mapAnalyticsFieldNames(provided)).toEqual(expected);
+    });
+
+    it('should sort corrected field names alphabetically', () => {
+        // Arrange
+        const provided = [
+            'mobilePhone',
+            'address.line1',
+            'customer.firstName',
+            'lastName',
+            'customer.email',
+            'city'
+        ];
+
+        const expected = [
+            'addressCity',
+            'addressLine1',
+            'email',
+            'firstName',
+            'lastName',
+            'mobilePhone'
+        ];
+
+        // Act & Assert
+        expect(mapAnalyticsFieldNames(provided)).toEqual(expected);
     });
 });